chore: Clean converter code and add doc
